Move QuickStartGuidesCodec.ts in codecs folder
